Moved DataSQRLRepo to main test harness and moved repository infrastructure to separate repository.
