+++ sqrl-examples/datasqrl/data/package.8_e-WN-FzfckOREZ85JY6pv6ktQ.json
@@ -1,1 +1,1 @@
-{"name":"datasqrl.examples.Nutshop","version":"1.0.0","latestStable":true,"license":"Apache-2.0","repository":"https://github.com/datasqrl/examples","homepage":"https://datasqrl.com","documentation":"https://datasqrl.com/tutorials","readme":"README.md","description":"Nutshop example data","file":"8_e-WN-FzfckOREZ85JY6pv6ktQ.zip","hash":"234dfljks3984ujhdwiughe","authorid":"2345xerjk234","submissionTime":"2022-05-19T02:16:39.553244Z","files":[{"filename":"orders.table.json","filesize":"301"},{"filename":"products.table.json","filesize":"458"},{"filename":"schema.yml","filesize":"1507"}]}
+{"name":"datasqrl.examples.Nutshop","version":"0.1.0","latestStable":true,"license":"Apache-2.0","repository":"https://github.com/datasqrl/examples","homepage":"https://datasqrl.com","documentation":"https://datasqrl.com/tutorials","readme":"README.md","description":"Nutshop example data","file":"8_e-WN-FzfckOREZ85JY6pv6ktQ.zip","hash":"234dfljks3984ujhdwiughe","authorid":"2345xerjk234","submissionTime":"2022-05-19T02:16:39.553244Z","files":[{"filename":"orders.table.json","filesize":"301"},{"filename":"products.table.json","filesize":"458"},{"filename":"schema.yml","filesize":"1507"}]}
