+++ sqrl-examples/datasqrl/versions.query.graphql
@@ -1,6 +1,6 @@
 {
 Package(name: "datasqrl.examples.Nutshop") {
-    versions(version: "1.0.0") {
+    versions(version: "0.1.0") {
         version
         file
         hash
