+++ sqrl-planner/src/test/java/com/datasqrl/util/data/DataSQRLRepo.java
@@ -1,21 +1,21 @@
 /*
  * Copyright (c) 2021, DataSQRL. All rights reserved. Use is subject to license terms.
  */
-package com.datasqrl.packager;
+package com.datasqrl.util.data;
 
 import com.datasqrl.util.TestDataset;
 import com.datasqrl.util.TestGraphQLSchema;
 import com.datasqrl.util.TestScript;
 
 import java.nio.file.Path;
+import java.util.List;
 import java.util.Set;
 
-public class DataSQRL implements TestDataset {
+public class DataSQRLRepo implements TestDataset {
 
   public static final Path BASE_PATH = Path.of("..", "sqrl-examples", "datasqrl");
 
-  public static final DataSQRL INSTANCE = new DataSQRL();
-
+  public static final DataSQRLRepo INSTANCE = new DataSQRLRepo();
 
   @Override
   public String getName() {
@@ -44,7 +44,7 @@ public class DataSQRL implements TestDataset {
 
   public TestScript getScript() {
     return TestScript.of(this, BASE_PATH.resolve("repo.sqrl"),
-        "packages", "submissions").build();
+        "package", "submission").graphQLSchemas(List.of(getGraphQL())).build();
   }
 
   public TestGraphQLSchema getGraphQL() {
